requirements.txt
Flask
requests
gunicorn
Werkzeug
pymongo
python-dotenv
